feat(c-api) Extract `global` into its own module.
