+++ lib/c-api/src/wasm_c_api/externals/mod.rs
@@ -1,0 +1,1 @@
+pub mod global;
