+++ lib/c-api/src/wasm_c_api/mod.rs
@@ -1,6 +1,7 @@
 //! entrypoints for the standard C API
 
 pub mod engine;
+pub mod externals;
 pub mod store;
 pub(crate) mod utils;
 #[cfg(feature = "wasi")]
@@ -18,10 +19,11 @@ use thiserror::Error;
 // https://github.com/rust-lang/rust/issues/57966
 use crate::c_try;
 use crate::ordered_resolver::OrderedResolver;
+use externals::global::wasm_global_t;
 use wasmer::{
-    ExportType, Extern, ExternType, Function, FunctionType, Global, GlobalType, ImportType,
-    Instance, Memory, MemoryType, Module, Mutability, Pages, RuntimeError, Store, Table, TableType,
-    Val, ValType,
+    ExportType, Extern, ExternType, Function, FunctionType, GlobalType, ImportType, Instance,
+    Memory, MemoryType, Module, Mutability, Pages, RuntimeError, Store, Table, TableType, Val,
+    ValType,
 };
 #[cfg(feature = "jit")]
 use wasmer_engine_jit::JIT;
@@ -721,63 +723,6 @@ pub unsafe extern "C" fn wasm_func_param_arity(func: &wasm_func_t) -> usize {
 #[no_mangle]
 pub unsafe extern "C" fn wasm_func_result_arity(func: &wasm_func_t) -> usize {
     func.inner.ty().results().len()
-}
-
-#[repr(C)]
-pub struct wasm_global_t {
-    // maybe needs to hold onto instance
-    inner: Global,
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn wasm_global_new(
-    store_ptr: Option<NonNull<wasm_store_t>>,
-    gt: &wasm_globaltype_t,
-    val: &wasm_val_t,
-) -> Option<Box<wasm_global_t>> {
-    let gt = gt.as_globaltype();
-    let wasm_val = val.try_into().ok()?;
-    let store_ptr: NonNull<Store> = store_ptr?.cast::<Store>();
-    let store = store_ptr.as_ref();
-    let global = if gt.mutability.is_mutable() {
-        Global::new_mut(store, wasm_val)
-    } else {
-        Global::new(store, wasm_val)
-    };
-
-    Some(Box::new(wasm_global_t { inner: global }))
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn wasm_global_delete(_global: Option<Box<wasm_global_t>>) {}
-
-// TODO: figure out if these should be deep or shallow copies
-#[no_mangle]
-pub unsafe extern "C" fn wasm_global_copy(wasm_global: &wasm_global_t) -> Box<wasm_global_t> {
-    // do shallow copy
-    Box::new(wasm_global_t {
-        inner: wasm_global.inner.clone(),
-    })
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn wasm_global_get(wasm_global: &wasm_global_t, out: &mut wasm_val_t) {
-    let value = wasm_global.inner.get();
-    *out = value.try_into().unwrap();
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn wasm_global_set(wasm_global: &mut wasm_global_t, val: &wasm_val_t) {
-    let value: Val = val.try_into().unwrap();
-    wasm_global.inner.set(value);
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn wasm_global_same(
-    wasm_global1: &wasm_global_t,
-    wasm_global2: &wasm_global_t,
-) -> bool {
-    wasm_global1.inner.same(&wasm_global2.inner)
 }
 
 #[repr(C)]
